refactor: separate to bubble-payload component file
